add a few bugfixes for private algorithm
For counters without laplace noise and E in the private
algorithm (similar to alpha in the non-private algorithm) equal to 0,
the private algorithm gives the same results as the private algorithm.
The thresholds might be off by one because the private algorithm doesn't
send initial proposals (it might be possible to do so, but I'm not sure
how this would impact privacy).

For now I've turned on laplace noise, which has been fixed, and used the
correct E, but for small school capacitie it looks like this makes the
private capacities negative, so I need to test once we can synthesize
data. The private counters now use floats for the counts to support
laplace noise.